+++ src/utils/cleanup.js
@@ -1,11 +1,12 @@
 let { join } = require('path')
-let { sync: rm } = require('rimraf')
+let { rmSync } = require('fs')
 
 // Best effort local artifact cleanup
 module.exports = function cleanup () {
   try {
     // Clean up temp dir from root proxy + fingerprint
-    rm(join(process.cwd(), '__ARC_TMP__'))
+    let tmp = join(process.cwd(), '__ARC_TMP__')
+    rmSync(tmp, { recursive: true, force: true })
   }
   catch (err) { null } // Swallow errors, we may have to bubble something else
 }